Move rest of integration tests into the gquic folder
