+++ integrationtests/gquic/data_manager_test.go
@@ -1,9 +1,39 @@
-package integrationtests
+package gquic_test
 
 import (
-	. "github.com/onsi/ginkgo"
+	"crypto/md5"
+	"math/rand"
+	"time"
+
+		. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+
 )
+
+type dataManager struct {
+	data []byte
+	md5  []byte
+}
+
+func (m *dataManager) GenerateData(len int) error {
+	m.data = make([]byte, len)
+	r := rand.New(rand.NewSource(int64(time.Now().Nanosecond())))
+	_, err := r.Read(m.data)
+	if err != nil {
+		return err
+	}
+	sum := md5.Sum(m.data)
+	m.md5 = sum[:]
+	return nil
+}
+
+func (m *dataManager) GetData() []byte {
+	return m.data
+}
+
+func (m *dataManager) GetMD5() []byte {
+	return m.md5
+}
 
 var _ = Describe("Data Manager", func() {
 	dm := dataManager{}
