+++ integrationtests/gquic/integration_test.go
@@ -1,4 +1,4 @@
-package integrationtests
+package gquic_test
 
 import (
 	"bytes"
