+++ integrationtests/gquic/random_rtt_test.go
@@ -1,8 +1,9 @@
-package integrationtests
+package gquic_test
 
 import (
 	"bytes"
 	"fmt"
+	"math/rand"
 	"os/exec"
 	"strconv"
 	"time"
@@ -17,19 +18,52 @@ import (
 	. "github.com/onsi/gomega/gexec"
 )
 
-var _ = Describe("non-zero RTT", func() {
+// get a random duration between min and max
+func getRandomDuration(min, max time.Duration) time.Duration {
+	return min + time.Duration(rand.Int63n(int64(max-min)))
+}
+
+var _ = Describe("Random Duration Generator", func() {
+	rand.Seed(time.Now().UnixNano())
+	It("gets a random RTT", func() {
+		var min time.Duration = time.Hour
+		var max time.Duration
+
+		var sum time.Duration
+		rep := 10000
+		for i := 0; i < rep; i++ {
+			val := getRandomDuration(100*time.Millisecond, 500*time.Millisecond)
+			sum += val
+			if val < min {
+				min = val
+			}
+			if val > max {
+				max = val
+			}
+		}
+		avg := sum / time.Duration(rep)
+		Expect(avg).To(BeNumerically("~", 300*time.Millisecond, 5*time.Millisecond))
+		Expect(min).To(BeNumerically(">=", 100*time.Millisecond))
+		Expect(min).To(BeNumerically("<", 105*time.Millisecond))
+		Expect(max).To(BeNumerically(">", 495*time.Millisecond))
+		Expect(max).To(BeNumerically("<=", 500*time.Millisecond))
+	})
+})
+
+var _ = Describe("Random RTT", func() {
 	BeforeEach(func() {
 		dataMan.GenerateData(dataLen)
 	})
 
 	var proxy *quicproxy.QuicProxy
 
-	runRTTTest := func(rtt time.Duration, version protocol.VersionNumber) {
+	runRTTTest := func(minRtt, maxRtt time.Duration, version protocol.VersionNumber) {
+		rand.Seed(time.Now().UnixNano())
 		var err error
 		proxy, err = quicproxy.NewQuicProxy("localhost:", quicproxy.Opts{
 			RemoteAddr: "localhost:" + testserver.Port(),
 			DelayPacket: func(_ quicproxy.Direction, _ protocol.PacketNumber) time.Duration {
-				return rtt / 2
+				return getRandomDuration(minRtt, maxRtt)
 			},
 		})
 		Expect(err).ToNot(HaveOccurred())
@@ -59,13 +93,9 @@ var _ = Describe("non-zero RTT", func() {
 		version := protocol.SupportedVersions[i]
 
 		Context(fmt.Sprintf("with quic version %d", version), func() {
-			roundTrips := [...]int{10, 50, 100, 200}
-			for _, rtt := range roundTrips {
-				It(fmt.Sprintf("gets a 500kB file with %dms RTT", rtt), func() {
-					dataMan.GenerateData(dataLen)
-					runRTTTest(time.Duration(rtt)*time.Millisecond, version)
-				})
-			}
+			It("gets a file a random RTT between 10ms and 30ms", func() {
+				runRTTTest(10*time.Millisecond, 30*time.Millisecond, version)
+			})
 		})
 	}
 })
