+++ integrationtests/gquic/rtt_test.go
@@ -1,9 +1,8 @@
-package integrationtests
+package gquic_test
 
 import (
 	"bytes"
 	"fmt"
-	"math/rand"
 	"os/exec"
 	"strconv"
 	"time"
@@ -18,52 +17,19 @@ import (
 	. "github.com/onsi/gomega/gexec"
 )
 
-// get a random duration between min and max
-func getRandomDuration(min, max time.Duration) time.Duration {
-	return min + time.Duration(rand.Int63n(int64(max-min)))
-}
-
-var _ = Describe("Random Duration Generator", func() {
-	rand.Seed(time.Now().UnixNano())
-	It("gets a random RTT", func() {
-		var min time.Duration = time.Hour
-		var max time.Duration
-
-		var sum time.Duration
-		rep := 10000
-		for i := 0; i < rep; i++ {
-			val := getRandomDuration(100*time.Millisecond, 500*time.Millisecond)
-			sum += val
-			if val < min {
-				min = val
-			}
-			if val > max {
-				max = val
-			}
-		}
-		avg := sum / time.Duration(rep)
-		Expect(avg).To(BeNumerically("~", 300*time.Millisecond, 5*time.Millisecond))
-		Expect(min).To(BeNumerically(">=", 100*time.Millisecond))
-		Expect(min).To(BeNumerically("<", 105*time.Millisecond))
-		Expect(max).To(BeNumerically(">", 495*time.Millisecond))
-		Expect(max).To(BeNumerically("<=", 500*time.Millisecond))
-	})
-})
-
-var _ = Describe("Random RTT", func() {
+var _ = Describe("non-zero RTT", func() {
 	BeforeEach(func() {
 		dataMan.GenerateData(dataLen)
 	})
 
 	var proxy *quicproxy.QuicProxy
 
-	runRTTTest := func(minRtt, maxRtt time.Duration, version protocol.VersionNumber) {
-		rand.Seed(time.Now().UnixNano())
+	runRTTTest := func(rtt time.Duration, version protocol.VersionNumber) {
 		var err error
 		proxy, err = quicproxy.NewQuicProxy("localhost:", quicproxy.Opts{
 			RemoteAddr: "localhost:" + testserver.Port(),
 			DelayPacket: func(_ quicproxy.Direction, _ protocol.PacketNumber) time.Duration {
-				return getRandomDuration(minRtt, maxRtt)
+				return rtt / 2
 			},
 		})
 		Expect(err).ToNot(HaveOccurred())
@@ -93,9 +59,13 @@ var _ = Describe("Random RTT", func() {
 		version := protocol.SupportedVersions[i]
 
 		Context(fmt.Sprintf("with quic version %d", version), func() {
-			It("gets a file a random RTT between 10ms and 30ms", func() {
-				runRTTTest(10*time.Millisecond, 30*time.Millisecond, version)
-			})
+			roundTrips := [...]int{10, 50, 100, 200}
+			for _, rtt := range roundTrips {
+				It(fmt.Sprintf("gets a 500kB file with %dms RTT", rtt), func() {
+					dataMan.GenerateData(dataLen)
+					runRTTTest(time.Duration(rtt)*time.Millisecond, version)
+				})
+			}
 		})
 	}
 })
